pyproject.toml
[tool.poetry]

name = "datafiles"
version = "1.1"
description = "File-based ORM for dataclasses."

license = "MIT"

authors = ["Jace Browning <jacebrowning@gmail.com>"]

readme = "README.md"

homepage = "https://pypi.org/project/datafiles"
documentation = "https://datafiles.readthedocs.io"
repository = "https://github.com/jacebrowning/datafiles"

keywords = [
    "dataclasses",
    "serialization",
    "type-annotations",
    "object-relational mapping",
    "YAML",
    "JSON",
    "TOML",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python",
    "Topic :: Software Development",
    "Topic :: Utilities",
]

[tool.poetry.dependencies]

python = "^3.7"

# Formats
"ruamel.yaml" = "^0.17.20"
tomlkit = "^0.7.2"

# ORM
parse = "^1.12"

# Utilities
cached_property = "^1.5"
classproperties = "^0.2"
minilog = "^2.0"

# Typing
typing-extensions = "^3.7"

[tool.poetry.dev-dependencies]

# Formatters
black = "^22.1"
isort = "^5.10"

# Linters
mypy = "~0.910"
pylint = "~2.7.4"
pydocstyle = "*"

# Testing
pytest = "^6.2.5"
pytest-describe = "^2.0"
pytest-expecter = "^2.1"
pytest-mock = "*"
pytest-random = "*"
pytest-repeat = "*"
pytest-watch = "*"
pytest-cov = "^2.8.1"
pytest-profiling = "*"

# Coverage
coveragespace = "^4.0"

# Documentation
mkdocs = "^1.2.3"
pygments = "^2.10"

# Notebooks
idna = "^3.3" # papermill dependency
ipython = "^7.31"
jupyter = "^1.0"
nbstripout ="~0.4"
papermill = "^2.3"

[tool.black]

quiet = true

[tool.isort]

profile = "black"

[build-system]

requires = ["poetry-core>=1.0"]
build-backend = "poetry.core.masonry.api"
